requirements.txt
torch>=2.0.0
gymnasium>=1.0.0
flappy-bird-gymnasium>=0.4.0
pygame>=2.6.0
PyYAML>=6.0
